ocr-worker/pyproject.toml
[project]
name = "ocr-worker"
version = "0.1.0"
description = "OCR worker service for PaperlessClone"
requires-python = ">=3.11"
dependencies = [
    "pika>=1.3.2",
    "minio>=7.2.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "pdf2image>=1.16.3",
    "pypdf2>=3.0.1",
    "google-generativeai>=0.8.3",
]
